style: moved SubmissionScore to /submissions component to colocate related files
